Just treat msmb_data package as a different data_home
